Fix makefile : flags, wildcard

+++ src/parser/parser_free.c
@@ -6,7 +6,7 @@
 /*   By: hsoysal <hsoysal@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/11 20:15:20 by hsoysal           #+#    #+#             */
-/*   Updated: 2025/06/11 20:19:52 by hsoysal          ###   ########.fr       */
+/*   Updated: 2025/06/20 14:32:44 by parden           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,19 +25,19 @@ static void	free_bonus(t_scene *scene)
 	int	i;
 
 	i = -1;
-	while (++i < scene->plane_count)
+	while ((size_t)++i < scene->plane_count)
 		free_damier_and_heightmap(scene->planes[i].damier,
 			scene->planes[i].heightmap);
 	i = -1;
-	while (++i < scene->sphere_count)
+	while ((size_t)++i < scene->sphere_count)
 		free_damier_and_heightmap(scene->spheres[i].damier,
 			scene->spheres[i].heightmap);
 	i = -1;
-	while (++i < scene->cylinder_count)
+	while ((size_t)++i < scene->cylinder_count)
 		free_damier_and_heightmap(scene->cylinders[i].damier,
 			scene->cylinders[i].heightmap);
 	i = -1;
-	while (++i < scene->cone_count)
+	while ((size_t)++i < scene->cone_count)
 		free_damier_and_heightmap(scene->cones[i].damier,
 			scene->cones[i].heightmap);
 }
